+++ src/main.c
@@ -7,8 +7,7 @@ int	main(int ac, char **av, char **envp)
 	(void)av;
 	if (ac > 1)
 		return (EXIT_FAILURE); // TODO (Ed) Deal with non-interactive mode
-//	char	*input = "\"'\"Hello\"'\"";
-	char	*input = "echo 'hello'\"world\"";
+	char	*input = "echo 'a''b'\"c\"";
 	cmd_lst = init_cmd_lst(input);
 	if (!cmd_lst)
 		return (EXIT_FAILURE);
