--- src/quote_handler.c
@@ -1,39 +1,0 @@
-#include "minishell.h"
-
-int	skip_quotes(char *input, int *index, char *unmatched_quote)
-{
-	char	quote;
-
-	quote = input[*index];
-	(*index)++;
-	while (input[*index] && input[*index] != quote)
-		(*index)++;
-	if (!input[*index])
-	{
-		*unmatched_quote = quote;
-		return (0);
-	}
-	(*index)++;
-	return (1);
-}
-
-char	*remove_quotes(char *str)
-{
-	size_t	len;
-	char	*cleaned;
-	char	first;
-	char	last;
-
-	len = ft_strlen(str);
-	if (len < 2)
-		return (str);
-	first = str[0];
-	last = str[len - 1];
-	if ((first == '\'' || first == '"') && first == last)
-	{
-		cleaned = ft_substr(str, 1, len - 2);
-		free(str);
-		return (cleaned);
-	}
-	return (str);
-}
