+++ src/utils/string_utils.c
